+++ src/event/common.rs
@@ -1765,6 +1765,7 @@ mod tests {
 
     fn port_scan_fields() -> PortScanFields {
         PortScanFields {
+            sensor: String::new(),
             src_addr: IpAddr::V4(Ipv4Addr::LOCALHOST),
             dst_addr: IpAddr::V4(Ipv4Addr::new(127, 0, 0, 2)),
             dst_ports: vec![80, 443, 8000, 8080, 8888, 8443, 9000, 9001, 9002],
@@ -1778,6 +1779,7 @@ mod tests {
 
     fn multi_host_port_scan_fields() -> MultiHostPortScanFields {
         MultiHostPortScanFields {
+            sensor: String::new(),
             src_addr: IpAddr::V4(Ipv4Addr::LOCALHOST),
             dst_addrs: vec![
                 IpAddr::V4(Ipv4Addr::new(127, 0, 0, 2)),
@@ -1794,6 +1796,7 @@ mod tests {
 
     fn external_ddos_fields() -> ExternalDdosFields {
         ExternalDdosFields {
+            sensor: String::new(),
             src_addrs: vec![
                 IpAddr::V4(Ipv4Addr::LOCALHOST),
                 IpAddr::V4(Ipv4Addr::new(127, 0, 0, 3)),
@@ -1836,6 +1839,7 @@ mod tests {
 
     fn ftp_brute_force_fields() -> FtpBruteForceFields {
         FtpBruteForceFields {
+            sensor: String::new(),
             src_addr: IpAddr::V4(Ipv4Addr::LOCALHOST),
             dst_addr: IpAddr::V4(Ipv4Addr::new(127, 0, 0, 2)),
             dst_port: 21,
@@ -1936,6 +1940,7 @@ mod tests {
 
     fn ldap_brute_force_fields() -> LdapBruteForceFields {
         LdapBruteForceFields {
+            sensor: String::new(),
             src_addr: IpAddr::V4(Ipv4Addr::LOCALHOST),
             dst_addr: IpAddr::V4(Ipv4Addr::new(127, 0, 0, 2)),
             dst_port: 389,
@@ -1953,6 +1958,7 @@ mod tests {
 
     fn rdp_brute_force_fields() -> RdpBruteForceFields {
         RdpBruteForceFields {
+            sensor: String::new(),
             src_addr: IpAddr::V4(Ipv4Addr::LOCALHOST),
             dst_addrs: vec![
                 IpAddr::V4(Ipv4Addr::new(127, 0, 0, 2)),