+++ src/event/conn.rs
@@ -41,8 +41,9 @@ impl PortScanFields {
     #[must_use]
     pub fn syslog_rfc5424(&self) -> String {
         format!(
-            "category={:?} src_addr={:?} dst_addr={:?} dst_ports={:?} start_time={:?} end_time={:?} proto={:?} confidence={:?}",
+            "category={:?} sensor={:?} src_addr={:?} dst_addr={:?} dst_ports={:?} start_time={:?} end_time={:?} proto={:?} confidence={:?}",
             self.category.to_string(),
+            self.sensor,
             self.src_addr.to_string(),
             self.dst_addr.to_string(),
             vector_to_string(&self.dst_ports),
@@ -56,6 +57,7 @@ impl PortScanFields {
 
 #[derive(Serialize, Deserialize)]
 pub struct PortScanFieldsV0_41 {
+    pub sensor: String,
     pub src_addr: IpAddr,
     pub dst_addr: IpAddr,
     pub dst_ports: Vec<u16>,
@@ -69,6 +71,7 @@ pub struct PortScanFieldsV0_41 {
 impl From<PortScanFieldsV0_39> for PortScanFieldsV0_41 {
     fn from(value: PortScanFieldsV0_39) -> Self {
         Self {
+            sensor: String::new(),
             src_addr: value.src_addr,
             dst_addr: value.dst_addr,
             dst_ports: value.dst_ports,
@@ -95,6 +98,7 @@ pub struct PortScanFieldsV0_39 {
 #[allow(clippy::module_name_repetitions)]
 #[derive(Serialize, Deserialize)]
 pub struct PortScan {
+    pub sensor: String,
     pub time: DateTime<Utc>,
     pub src_addr: IpAddr,
     pub dst_addr: IpAddr,
@@ -126,6 +130,7 @@ impl fmt::Display for PortScan {
 impl PortScan {
     pub(super) fn new(time: DateTime<Utc>, fields: &PortScanFields) -> Self {
         PortScan {
+            sensor: fields.sensor.clone(),
             time,
             src_addr: fields.src_addr,
             dst_addr: fields.dst_addr,
@@ -208,8 +213,9 @@ impl MultiHostPortScanFields {
     #[must_use]
     pub fn syslog_rfc5424(&self) -> String {
         format!(
-            "category={:?} src_addr={:?} dst_addrs={:?} dst_port={:?} proto={:?} start_time={:?} end_time={:?} confidence={:?}",
+            "category={:?} sensor={:?} src_addr={:?} dst_addrs={:?} dst_port={:?} proto={:?} start_time={:?} end_time={:?} confidence={:?}",
             self.category.to_string(),
+            self.sensor,
             self.src_addr.to_string(),
             vector_to_string(&self.dst_addrs),
             self.dst_port.to_string(),
@@ -223,6 +229,7 @@ impl MultiHostPortScanFields {
 
 #[derive(Serialize, Deserialize)]
 pub struct MultiHostPortScanFieldsV0_41 {
+    pub sensor: String,
     pub src_addr: IpAddr,
     pub dst_port: u16,
     pub dst_addrs: Vec<IpAddr>,
@@ -236,6 +243,7 @@ pub struct MultiHostPortScanFieldsV0_41 {
 impl From<MultiHostPortScanFieldsV0_39> for MultiHostPortScanFieldsV0_41 {
     fn from(value: MultiHostPortScanFieldsV0_39) -> Self {
         Self {
+            sensor: String::new(),
             src_addr: value.src_addr,
             dst_port: value.dst_port,
             dst_addrs: value.dst_addrs,
@@ -262,6 +270,7 @@ pub struct MultiHostPortScanFieldsV0_39 {
 #[allow(clippy::module_name_repetitions)]
 #[derive(Serialize, Deserialize)]
 pub struct MultiHostPortScan {
+    pub sensor: String,
     pub time: DateTime<Utc>,
     pub src_addr: IpAddr,
     pub dst_port: u16,
@@ -293,6 +302,7 @@ impl fmt::Display for MultiHostPortScan {
 impl MultiHostPortScan {
     pub(super) fn new(time: DateTime<Utc>, fields: &MultiHostPortScanFields) -> Self {
         MultiHostPortScan {
+            sensor: fields.sensor.clone(),
             time,
             src_addr: fields.src_addr,
             dst_port: fields.dst_port,
@@ -373,8 +383,9 @@ impl ExternalDdosFields {
     #[must_use]
     pub fn syslog_rfc5424(&self) -> String {
         format!(
-            "category={:?} src_addrs={:?} dst_addr={:?} proto={:?} start_time={:?} end_time={:?} confidence={:?}",
+            "category={:?} sensor={:?} src_addrs={:?} dst_addr={:?} proto={:?} start_time={:?} end_time={:?} confidence={:?}",
             self.category.to_string(),
+            self.sensor,
             vector_to_string(&self.src_addrs),
             self.dst_addr.to_string(),
             self.proto.to_string(),
@@ -387,6 +398,7 @@ impl ExternalDdosFields {
 
 #[derive(Serialize, Deserialize)]
 pub struct ExternalDdosFieldsV0_41 {
+    pub sensor: String,
     pub src_addrs: Vec<IpAddr>,
     pub dst_addr: IpAddr,
     pub proto: u8,
@@ -399,6 +411,7 @@ pub struct ExternalDdosFieldsV0_41 {
 impl From<ExternalDdosFieldsV0_39> for ExternalDdosFieldsV0_41 {
     fn from(value: ExternalDdosFieldsV0_39) -> Self {
         Self {
+            sensor: String::new(),
             src_addrs: value.src_addrs,
             dst_addr: value.dst_addr,
             proto: value.proto,
@@ -423,6 +436,7 @@ pub struct ExternalDdosFieldsV0_39 {
 #[allow(clippy::module_name_repetitions)]
 #[derive(Serialize, Deserialize)]
 pub struct ExternalDdos {
+    pub sensor: String,
     pub time: DateTime<Utc>,
     pub src_addrs: Vec<IpAddr>,
     pub dst_addr: IpAddr,
@@ -452,6 +466,7 @@ impl fmt::Display for ExternalDdos {
 impl ExternalDdos {
     pub(super) fn new(time: DateTime<Utc>, fields: &ExternalDdosFields) -> Self {
         ExternalDdos {
+            sensor: fields.sensor.clone(),
             time,
             src_addrs: fields.src_addrs.clone(),
             dst_addr: fields.dst_addr,