+++ src/event/ftp.rs
@@ -43,8 +43,9 @@ impl FtpBruteForceFields {
     #[must_use]
     pub fn syslog_rfc5424(&self) -> String {
         format!(
-            "category={:?} src_addr={:?} dst_addr={:?} dst_port={:?} proto={:?} user_list={:?} start_time={:?} end_time={:?} is_internal={:?} confidence={:?}",
+            "category={:?} sensor={:?} src_addr={:?} dst_addr={:?} dst_port={:?} proto={:?} user_list={:?} start_time={:?} end_time={:?} is_internal={:?} confidence={:?}",
             self.category.to_string(),
+            self.sensor,
             self.src_addr.to_string(),
             self.dst_addr.to_string(),
             self.dst_port.to_string(),
@@ -60,6 +61,7 @@ impl FtpBruteForceFields {
 
 #[derive(Serialize, Deserialize)]
 pub struct FtpBruteForceFieldsV0_41 {
+    pub sensor: String,
     pub src_addr: IpAddr,
     pub dst_addr: IpAddr,
     pub dst_port: u16,
@@ -75,6 +77,7 @@ pub struct FtpBruteForceFieldsV0_41 {
 impl From<FtpBruteForceFieldsV0_39> for FtpBruteForceFieldsV0_41 {
     fn from(value: FtpBruteForceFieldsV0_39) -> Self {
         Self {
+            sensor: String::new(),
             src_addr: value.src_addr,
             dst_addr: value.dst_addr,
             dst_port: value.dst_port,
@@ -104,6 +107,7 @@ pub struct FtpBruteForceFieldsV0_39 {
 
 #[derive(Serialize, Deserialize)]
 pub struct FtpBruteForce {
+    pub sensor: String,
     pub time: DateTime<Utc>,
     pub src_addr: IpAddr,
     pub dst_addr: IpAddr,
@@ -139,6 +143,7 @@ impl fmt::Display for FtpBruteForce {
 impl FtpBruteForce {
     pub(super) fn new(time: DateTime<Utc>, fields: &FtpBruteForceFields) -> Self {
         FtpBruteForce {
+            sensor: fields.sensor.clone(),
             time,
             src_addr: fields.src_addr,
             dst_addr: fields.dst_addr,