+++ src/event/rdp.rs
@@ -36,8 +36,9 @@ impl RdpBruteForceFields {
     #[must_use]
     pub fn syslog_rfc5424(&self) -> String {
         format!(
-            "category={:?} src_addr={:?} dst_addrs={:?} start_time={:?} end_time={:?} proto={:?} confidence={:?}",
+            "category={:?} sensor={:?} src_addr={:?} dst_addrs={:?} start_time={:?} end_time={:?} proto={:?} confidence={:?}",
             self.category.to_string(),
+            self.sensor,
             self.src_addr.to_string(),
             vector_to_string(&self.dst_addrs),
             self.start_time.to_rfc3339(),
@@ -50,6 +51,7 @@ impl RdpBruteForceFields {
 
 #[derive(Serialize, Deserialize)]
 pub struct RdpBruteForceFieldsV0_41 {
+    pub sensor: String,
     pub src_addr: IpAddr,
     pub dst_addrs: Vec<IpAddr>,
     pub start_time: DateTime<Utc>,
@@ -62,6 +64,7 @@ pub struct RdpBruteForceFieldsV0_41 {
 impl From<RdpBruteForceFieldsV0_39> for RdpBruteForceFieldsV0_41 {
     fn from(value: RdpBruteForceFieldsV0_39) -> Self {
         Self {
+            sensor: String::new(),
             src_addr: value.src_addr,
             dst_addrs: value.dst_addrs,
             start_time: value.start_time,
@@ -85,6 +88,7 @@ pub struct RdpBruteForceFieldsV0_39 {
 
 #[derive(Serialize, Deserialize)]
 pub struct RdpBruteForce {
+    pub sensor: String,
     pub time: DateTime<Utc>,
     pub src_addr: IpAddr,
     pub dst_addrs: Vec<IpAddr>,
@@ -114,6 +118,7 @@ impl fmt::Display for RdpBruteForce {
 impl RdpBruteForce {
     pub(super) fn new(time: DateTime<Utc>, fields: &RdpBruteForceFields) -> Self {
         RdpBruteForce {
+            sensor: fields.sensor.clone(),
             time,
             src_addr: fields.src_addr,
             dst_addrs: fields.dst_addrs.clone(),